buildkite search: rename and move
