+++ misc/python/materialize/buildkite_insights/annotation_search/README.md
@@ -1,0 +1,3 @@
+# Buildkite Annotation Search
+
+Script to search annotations for (error) messages.
