+++ misc/python/materialize/buildkite_insights/annotation_search/search.py
@@ -23,12 +23,12 @@ from materialize.buildkite_insights.cache.cache_constants import (
     FETCH_MODE_CHOICES,
     FetchMode,
 )
-from materialize.buildkite_insights.failure_search.search_result_presentation import (
+from materialize.buildkite_insights.annotation_search.search_result_presentation import (
     print_before_search_results,
     print_match,
     print_summary,
 )
-from materialize.buildkite_insights.failure_search.search_utility import (
+from materialize.buildkite_insights.annotation_search.search_utility import (
     _search_value_to_pattern,
 )
