--- misc/python/materialize/buildkite_insights/failure_search/README.md
@@ -1,3 +1,0 @@
-# Buildkite Failure Search
-
-Script to search annotations for error messages.
